MOD: Converting kilometers to miles

+++ ConvertToMiles.java
@@ -1,7 +1,7 @@
 public class ConvertToMiles{
 	public static void main(String[] args){
-		float kilometers = 10.8;
-		float oneMile = 1.6;
+		float kilometers = (float)10.8;
+		float oneMile = (float)1.6;
 		float miles = kilometers*oneMile;
 		System.out.print("The distance 10.8 km in miles is " + miles);
 	}
